Move organization_uuid validation to permissions
