+++ location/permissions.py
@@ -1,5 +1,14 @@
+import re
+
 from rest_framework.permissions import IsAuthenticated
-from rest_framework.exceptions import PermissionDenied
+from rest_framework.exceptions import PermissionDenied, ValidationError
+
+
+def _valid_uuid4(uuid_string):
+    uuid4hex = re.compile('^[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}\Z',
+                          re.I)
+    match = uuid4hex.match(uuid_string)
+    return bool(match)
 
 
 class AllowOptionsAuthentication(IsAuthenticated):
@@ -7,8 +16,12 @@ class AllowOptionsAuthentication(IsAuthenticated):
         if request.method == 'OPTIONS':
             return True
 
-        if getattr(request, 'session', None) and \
-                request.session.get('jwt_organization_uuid'):
+        if getattr(request, 'session', None) and request.session.get('jwt_organization_uuid'):
+            organization_uuid = request.session['jwt_organization_uuid']
+            if not _valid_uuid4(organization_uuid):
+                raise ValidationError(
+                    f'organization_uuid from JWT Token "{organization_uuid}" is not a valid UUID.'
+                )
             return True
         return False
